sprint17/requirements.txt
streamlit==1.33.0
streamlit-drawable-canvas==0.9.3
onnxruntime==1.19.2
opencv-python-headless==4.10.0.84
Pillow==10.4.0
numpy==1.26.4
pandas==2.2.2
requests==2.32.3
